add .env file

+++ src/database/knex/migrations/20230404134243_productsTable.js
@@ -2,7 +2,7 @@ exports.up = (knex) =>
   knex.schema.createTable('products', (table) => {
     table.integer('id').notNullable().primary();
     table.text('name').notNullable();
-    table.integer('price').notNullable();
+    table.decimal('price').notNullable();
     table.integer('stock').notNullable();
   });
